[FC-36891] fix varnish monitoring
The default monitoring check relied on varnish to run on a specific
address.
Since this address can be changed, the Sensu check needs to be adjusted
to correctly infer the host and port that varnish is running on

The listen address can only be reliably queried at runtime
(without reimplementing Varnishd's command line parsing logic in nix)
hence why multiple checks - one per address - are not possible.