fc-postgresql: support upgrades with custom extensions
FC-42163

Upgrading postgresql to a new major with fc-postgresql fails when using
custom extensions, defined via `services.postgresql.extraPlugins`.

This patch modifies the relevant agent code to use a postgresql package
WITH all needed plugins for upgrades:

* for automatic upgrades, these extensions are discovered automatically.
* for manual upgrades, the user must specify those manually with
  `--extension-names`.

The original use-case of a customer was to upgrade postgresql to a newer
version with `postgresqlPackages.anonymizer` installed (FC-42089). This
is a little trickier since it requires `anon` to be listed in
`shared_preload_libraries`. For this, all `shared_preload_libraries` are
now extracted from `postgresql.conf` and passed to `pg_upgrade` via CLI.
This code is based on the assumption that the old `postgresql.conf` is
generated by Nix and thus ignores all edge-cases, the postgresql config
format may have.